refactor: moving code around
