+++ src/Pages/+Shared/Layout/Main.tsx
@@ -1,6 +1,6 @@
-import React, { FC } from 'react';
-import { StyledContainer } from './Main.styled';
-import { Navigation } from './Nav';
+import React, { FC } from "react";
+import { StyledContainer } from "./Main.styled";
+import { Navigation } from "../Nav";
 
 export const MainLayout: FC<{}> = ({ children }) => (
   <StyledContainer>
