+++ src/Pages/+Shared/Layout/index.ts
@@ -1,0 +1,2 @@
+export * from "./Main";
+export * from "./Main.styled";
