+++ src/Pages/+Shared/Nav/index.ts
@@ -1,0 +1,1 @@
+export * from "../Nav/Nav";
