+++ src/Pages/Home/Home.tsx
@@ -1,4 +1,6 @@
-import React from 'react';
-import { MainLayout } from '../../Shared/Layout/Main';
+import React from "react";
+
+// TODO: Should be absolute path since outside of futur lazy load HOME page
+import { MainLayout } from "../+Shared/Layout/Main";
 
 export const HomePage = () => <MainLayout>Home Page</MainLayout>;
