--- src/Shared/Layout/index.ts
@@ -1,3 +1,0 @@
-export * from './Main';
-export * from './Main.styled';
-export * from './Nav';
